+++ prover/src/lib.rs
@@ -38,8 +38,8 @@ pub use miden_air::{
     DeserializationError, ExecutionProof, FieldExtension, HashFunction, ProvingOptions,
 };
 pub use miden_processor::{
-    AdviceInputs, AsyncHost, BaseHost, ExecutionError, InputError, StackInputs, StackOutputs,
-    SyncHost, Word, crypto, math, utils,
+    AdviceInputs, AsyncHost, BaseHost, ExecutionError, InputError, PrecompileRequest, StackInputs,
+    StackOutputs, SyncHost, Word, crypto, math, utils,
 };
 pub use winter_prover::{Proof, crypto::MerkleTree as MerkleTreeVC};
 
@@ -87,6 +87,9 @@ pub fn prove(
 
     let stack_outputs = trace.stack_outputs().clone();
     let hash_fn = options.hash_fn();
+
+    // extract precompile requests from the trace to include in the proof
+    let precompile_requests = trace.precompile_requests().to_vec();
 
     // generate STARK proof
     let proof = match hash_fn {
@@ -136,7 +139,8 @@ pub fn prove(
         },
     }
     .map_err(ExecutionError::ProverError)?;
-    let proof = ExecutionProof::new(proof, hash_fn);
+
+    let proof = ExecutionProof::new(proof, hash_fn, precompile_requests);
 
     Ok((stack_outputs, proof))
 }