+++ stdlib/tests/crypto/stark/mod.rs
@@ -91,7 +91,7 @@ pub fn generate_recursive_verifier_data(
 
     // build public inputs and generate the advice data needed for recursive proof verification
     let pub_inputs = PublicInputs::new(program_info, stack_inputs, stack_outputs);
-    let (_, proof) = proof.into_parts();
+    let (_, proof, _precompile_requests) = proof.into_parts();
     Ok(generate_advice_inputs(proof, pub_inputs).unwrap())
 }